+++ src/aqua/__init__.py
@@ -5,6 +5,7 @@ from .graphics import plot_hovmoller
 from .lra_generator import LRAgenerator
 from .reader import Reader, catalog, Streaming, inspect_catalog
 from. regridder import Regridder
+from .fldstat import FldStat
 from .fixer import Fixer
 from .accessor import AquaAccessor
 
@@ -12,5 +13,6 @@ __all__ = ["plot_single_map", "plot_maps", "plot_single_map_diff", "plot_timeser
            "plot_hovmoller",
            "LRAgenerator",
            "Reader", "catalog", "Streaming", "inspect_catalog",
-            "Fixer",
+           "Regridder", "FldStat",
+           "Fixer",
            "Regridder"]